requirements.txt
djangorestframework
markdown
django
django-filter
django-guardian
Pygments
PyYAML
uritemplate
